+++ lint-fs.yaml
@@ -20,7 +20,7 @@ rules:
   - ^./package.json$
   - ^./run-tests.js$
   - ^./stryker.conf.json$
-  - ^.(/[a-z]+((\d)|([A-Z0-9][a-z0-9]+))*([A-Z])?)*/([a-z]+((\d)|([A-Z0-9][a-z0-9]+))*([A-Z])?)/\6(.test)?.js$
+  - ^.(/[a-z]+((\d)|([A-Z0-9][a-z0-9]+))*([A-Z])?)*/([a-z]+((\d)|([A-Z0-9][a-z0-9]+))*([A-Z])?)/(\6(.test)?|helpers).js$
 
 # mode: verbose | silent
 # провалидировать, что обязательно должно быть либо verbose, либо silent
